se agrega el helper con la logica de negocio

+++ models/estudiante.go
@@ -2,8 +2,8 @@ package models
 
 type Estudiante struct{
 	id 					int
-	uqCodgigoEstudiante int
-	codMateria 			int
+	uqCodgigoEstudiante float64
+	codMateria 			float64
 	nota1 				float64
 	nota2 				float64
 	nota3 				float64
